fix wind scaling problem in prsflg==2
In mod_random_forcing, the prsflg==2 option scales the wind perturbation
magnitude to that specified in the namelist. The scaling analysis use
the pressure perturbation magnitude and the pressure wind relationship
to compute the expected wind magnitude. But the grid spacing
used in the code is minscpx which is not spatially varying. Grid spacing
is varying (scpx and scpy) in the computation of pressure gradients
dpresx and dpresy, therefore changed to using varying grid spacing in
wprsfac computation, so that resulting wind perturbation magnitude won't
be affected by varying dx.